Integrate runtime validation
