+++ web-common/src/features/charts/editor/ChartsEditor.svelte
@@ -2,11 +2,14 @@
   import type { EditorView } from "@codemirror/view";
   import YAMLEditor from "@rilldata/web-common/components/editor/YAMLEditor.svelte";
   import ChartsEditorContainer from "@rilldata/web-common/features/charts/editor/ChartsEditorContainer.svelte";
+  import { useChart } from "@rilldata/web-common/features/charts/selectors";
   import {
     getFileAPIPathFromNameAndType,
     getFilePathFromNameAndType,
   } from "@rilldata/web-common/features/entity-management/entity-mappers";
+  import { getAllErrorsForFile } from "@rilldata/web-common/features/entity-management/resources-store";
   import { EntityType } from "@rilldata/web-common/features/entity-management/types";
+  import { mapParseErrorsToLines } from "@rilldata/web-common/features/metrics-views/errors";
   import { createPlaceholder } from "@rilldata/web-common/features/metrics-views/workspace/editor/create-placeholder";
   import { debounce } from "@rilldata/web-common/lib/create-debouncer";
   import {
@@ -14,6 +17,7 @@
     createRuntimeServicePutFile,
   } from "@rilldata/web-common/runtime-client";
   import { runtime } from "@rilldata/web-common/runtime-client/runtime-store";
+  import { useQueryClient } from "@tanstack/svelte-query";
 
   export let chartName: string;
 
@@ -22,6 +26,8 @@
 
   let view: EditorView;
   let editor: YAMLEditor;
+
+  $: chart = useChart($runtime.instanceId, chartName);
 
   /** create placeholder codemirror extension and set the inner svelte component
    * generated by the createPlaceholder callback.
@@ -36,6 +42,22 @@
 
   // get the yaml blob from the file.
   $: yaml = $fileQuery.data?.blob || "";
+
+  const queryClient = useQueryClient();
+  $: allErrors = getAllErrorsForFile(
+    queryClient,
+    $runtime.instanceId,
+    `/charts/${chartName}.yaml`,
+  );
+
+  $: lineBasedRuntimeErrors = mapParseErrorsToLines($allErrors, yaml);
+  /** display the main error (the first in this array) at the bottom */
+  $: mainError = [
+    ...lineBasedRuntimeErrors,
+    ...($chart.data?.meta?.reconcileError
+      ? [$chart.data?.meta?.reconcileError as any] // TODO: revisit error
+      : []),
+  ]?.at(0);
 
   async function updateChart(content: string) {
     try {
@@ -53,7 +75,7 @@
   const debounceUpdateChartContent = debounce(updateChart, QUERY_DEBOUNCE_TIME);
 </script>
 
-<ChartsEditorContainer error={undefined}>
+<ChartsEditorContainer error={yaml?.length ? mainError : undefined}>
   <YAMLEditor
     bind:this={editor}
     bind:view
